+++ src/Services/Time.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace OpenEuropa\TaskRunner\Services;
+
+/**
+ * Class Time
+ *
+ * @package OpenEuropa\TaskRunner\Services
+ */
+class Time
+{
+    /**
+     * @return int
+     */
+    public function getTimestamp()
+    {
+        return time();
+    }
+}